refactor(DatePicker): Move to `partials`
