+++ packages/react-component-library/src/components/DatePicker/partials/StyledInputWrapper.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+export const StyledInputWrapper = styled.div`
+  position: relative;
+  flex-grow: 1;
+  order: 2;
+`
