+++ packages/react-component-library/src/components/DatePicker/partials/StyledSeparator.tsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components'
+import { selectors } from '@royalnavy/design-tokens'
+
+const { color, spacing } = selectors
+
+export const StyledSeparator = styled.div`
+  width: 1px;
+  height: 100%;
+  margin-right: ${spacing('8')};
+  background: ${color('neutral', '100')};
+`
